requirements.dev.txt
astroid==1.6.3
flake8==3.5.0
pylint==1.8.4
pylint-django==0.11
coverage==4.5.1
sphinx==1.7.3
pytest-django==3.2.1
allure-pytest==2.3.2b1